+++ frontend/src/components/LoadMore/index.ts
@@ -1,2 +1,3 @@
 export { LoadMore } from './LoadMore'
 export { DeviceLoadMore } from './DeviceLoadMore'
+export { JobLoadMore } from './JobLoadMore'
